refactor(ui): centralize DA proxy auth, remove dead gateway routes, default to agui
- Extract da-proxy.ts with authenticateRequest, buildBackendHeaders,
  proxyJSONRequest, proxySSEStream helpers
- Rewrite files/content/interrupt-state/clear/suggest proxy routes to
  use centralized auth (fixes missing X-User-Context header)
- Remove Bearer headers from browser file fetches in
  DynamicAgentChatPanel (session cookies suffice for same-origin)
- Delete dead unified gateway routes (/api/chat/conversations/[id]/stream/
  and /invoke/)
- Change default agentProtocol from custom to agui
