+++ ui/src/app/api/dynamic-agents/conversations/[id]/clear/route.ts
@@ -9,74 +9,41 @@
 
 import { NextRequest, NextResponse } from "next/server";
 import {
-  withAuth,
-  withErrorHandler,
-  ApiError,
-} from "@/lib/api-middleware";
-import { getServerConfig } from "@/lib/config";
+  authenticateRequest,
+  getDynamicAgentsConfig,
+  proxyRequest,
+} from "@/lib/da-proxy";
 
 /**
  * POST /api/dynamic-agents/conversations/[id]/clear
  * Proxy to Dynamic Agents service to clear checkpoint data (admin only).
  */
-export const POST = withErrorHandler(
-  async (
-    request: NextRequest,
-    context: { params: Promise<{ id: string }> }
-  ) => {
-    const { id: conversationId } = await context.params;
+export async function POST(
+  request: NextRequest,
+  context: { params: Promise<{ id: string }> },
+): Promise<Response> {
+  const { id: conversationId } = await context.params;
 
-    if (!conversationId) {
-      throw new ApiError("Conversation ID is required", 400);
-    }
-
-    return await withAuth(request, async (req, user, session) => {
-      const config = getServerConfig();
-
-      if (!config.dynamicAgentsEnabled) {
-        throw new ApiError("Dynamic agents are not enabled", 403);
-      }
-
-      if (!config.dynamicAgentsUrl) {
-        throw new ApiError("Dynamic agents URL not configured", 500);
-      }
-
-      // Build the backend URL
-      const backendUrl = new URL(
-        `/api/v1/conversations/${conversationId}/clear`,
-        config.dynamicAgentsUrl
-      );
-
-      // Build headers for the backend request
-      const backendHeaders: Record<string, string> = {
-        "Content-Type": "application/json",
-      };
-
-      // Forward the access token to the Dynamic Agents backend
-      if (session.accessToken) {
-        backendHeaders["Authorization"] = `Bearer ${session.accessToken}`;
-      }
-
-      // Forward the request to the Dynamic Agents service
-      const response = await fetch(backendUrl.toString(), {
-        method: "POST",
-        headers: backendHeaders,
-      });
-
-      if (!response.ok) {
-        const errorText = await response.text();
-        let errorMessage = `Failed to clear conversation: ${response.status}`;
-        try {
-          const errorJson = JSON.parse(errorText);
-          errorMessage = errorJson.detail || errorMessage;
-        } catch {
-          // Use default error message
-        }
-        throw new ApiError(errorMessage, response.status);
-      }
-
-      const data = await response.json();
-      return NextResponse.json(data);
-    });
+  if (!conversationId) {
+    return NextResponse.json(
+      { success: false, error: "Conversation ID is required" },
+      { status: 400 },
+    );
   }
-);
+
+  // Authenticate
+  const authResult = await authenticateRequest(request);
+  if (authResult instanceof NextResponse) return authResult;
+
+  // Check DA config
+  const daConfig = getDynamicAgentsConfig();
+  if (daConfig instanceof NextResponse) return daConfig;
+
+  // Build backend URL
+  const backendUrl = new URL(
+    `/api/v1/conversations/${conversationId}/clear`,
+    daConfig.dynamicAgentsUrl,
+  );
+
+  return proxyRequest(backendUrl.toString(), "POST", authResult, "[clear]");
+}
