+++ ui/src/app/api/dynamic-agents/conversations/[id]/files/content/route.ts
@@ -10,168 +10,88 @@
 
 import { NextRequest, NextResponse } from "next/server";
 import {
-  withAuth,
-  withErrorHandler,
-  ApiError,
-} from "@/lib/api-middleware";
-import { getServerConfig } from "@/lib/config";
+  authenticateRequest,
+  getDynamicAgentsConfig,
+  proxyRequest,
+} from "@/lib/da-proxy";
+
+/**
+ * Validate common params and return backend URL, or a NextResponse error.
+ */
+async function resolveBackendUrl(
+  request: NextRequest,
+  context: { params: Promise<{ id: string }> },
+): Promise<{ backendUrl: string; authResult: Exclude<Awaited<ReturnType<typeof authenticateRequest>>, NextResponse> } | NextResponse> {
+  const { id: conversationId } = await context.params;
+
+  if (!conversationId) {
+    return NextResponse.json(
+      { success: false, error: "Conversation ID is required" },
+      { status: 400 },
+    );
+  }
+
+  const { searchParams } = new URL(request.url);
+  const agentId = searchParams.get("agent_id");
+  const path = searchParams.get("path");
+
+  if (!agentId) {
+    return NextResponse.json(
+      { success: false, error: "agent_id query parameter is required" },
+      { status: 400 },
+    );
+  }
+
+  if (!path) {
+    return NextResponse.json(
+      { success: false, error: "path query parameter is required" },
+      { status: 400 },
+    );
+  }
+
+  // Authenticate
+  const authResult = await authenticateRequest(request);
+  if (authResult instanceof NextResponse) return authResult;
+
+  // Check DA config
+  const daConfig = getDynamicAgentsConfig();
+  if (daConfig instanceof NextResponse) return daConfig;
+
+  // Build backend URL
+  const backendUrl = new URL(
+    `/api/v1/conversations/${conversationId}/files/content`,
+    daConfig.dynamicAgentsUrl,
+  );
+  backendUrl.searchParams.set("agent_id", agentId);
+  backendUrl.searchParams.set("path", path);
+
+  return { backendUrl: backendUrl.toString(), authResult };
+}
 
 /**
  * GET /api/dynamic-agents/conversations/[id]/files/content
  * Proxy to Dynamic Agents service to get file content from checkpointer.
  */
-export const GET = withErrorHandler(
-  async (
-    request: NextRequest,
-    context: { params: Promise<{ id: string }> }
-  ) => {
-    const { id: conversationId } = await context.params;
+export async function GET(
+  request: NextRequest,
+  context: { params: Promise<{ id: string }> },
+): Promise<Response> {
+  const resolved = await resolveBackendUrl(request, context);
+  if (resolved instanceof NextResponse) return resolved;
 
-    if (!conversationId) {
-      throw new ApiError("Conversation ID is required", 400);
-    }
-
-    const { searchParams } = new URL(request.url);
-    const agentId = searchParams.get("agent_id");
-    const path = searchParams.get("path");
-
-    if (!agentId) {
-      throw new ApiError("agent_id query parameter is required", 400);
-    }
-
-    if (!path) {
-      throw new ApiError("path query parameter is required", 400);
-    }
-
-    return await withAuth(request, async (req, user, session) => {
-      const config = getServerConfig();
-      
-      if (!config.dynamicAgentsEnabled) {
-        throw new ApiError("Dynamic agents are not enabled", 403);
-      }
-      
-      if (!config.dynamicAgentsUrl) {
-        throw new ApiError("Dynamic agents URL not configured", 500);
-      }
-      
-      // Build the backend URL
-      const backendUrl = new URL(
-        `/api/v1/conversations/${conversationId}/files/content`,
-        config.dynamicAgentsUrl
-      );
-      backendUrl.searchParams.set("agent_id", agentId);
-      backendUrl.searchParams.set("path", path);
-
-      // Build headers for the backend request
-      const backendHeaders: Record<string, string> = {
-        "Content-Type": "application/json",
-      };
-
-      // Forward the access token to the Dynamic Agents backend
-      if (session.accessToken) {
-        backendHeaders["Authorization"] = `Bearer ${session.accessToken}`;
-      }
-
-      // Forward the request to the Dynamic Agents service
-      const response = await fetch(backendUrl.toString(), {
-        method: "GET",
-        headers: backendHeaders,
-      });
-
-      if (!response.ok) {
-        const errorText = await response.text();
-        let errorMessage = `Failed to fetch file content: ${response.status}`;
-        try {
-          const errorJson = JSON.parse(errorText);
-          errorMessage = errorJson.detail || errorMessage;
-        } catch {
-          // Use default error message
-        }
-        throw new ApiError(errorMessage, response.status);
-      }
-
-      const data = await response.json();
-      return NextResponse.json(data);
-    });
-  }
-);
+  return proxyRequest(resolved.backendUrl, "GET", resolved.authResult, "[files/content]");
+}
 
 /**
  * DELETE /api/dynamic-agents/conversations/[id]/files/content
  * Proxy to Dynamic Agents service to delete a file from checkpointer.
  */
-export const DELETE = withErrorHandler(
-  async (
-    request: NextRequest,
-    context: { params: Promise<{ id: string }> }
-  ) => {
-    const { id: conversationId } = await context.params;
+export async function DELETE(
+  request: NextRequest,
+  context: { params: Promise<{ id: string }> },
+): Promise<Response> {
+  const resolved = await resolveBackendUrl(request, context);
+  if (resolved instanceof NextResponse) return resolved;
 
-    if (!conversationId) {
-      throw new ApiError("Conversation ID is required", 400);
-    }
-
-    const { searchParams } = new URL(request.url);
-    const agentId = searchParams.get("agent_id");
-    const path = searchParams.get("path");
-
-    if (!agentId) {
-      throw new ApiError("agent_id query parameter is required", 400);
-    }
-
-    if (!path) {
-      throw new ApiError("path query parameter is required", 400);
-    }
-
-    return await withAuth(request, async (req, user, session) => {
-      const config = getServerConfig();
-
-      if (!config.dynamicAgentsEnabled) {
-        throw new ApiError("Dynamic agents are not enabled", 403);
-      }
-
-      if (!config.dynamicAgentsUrl) {
-        throw new ApiError("Dynamic agents URL not configured", 500);
-      }
-
-      // Build the backend URL
-      const backendUrl = new URL(
-        `/api/v1/conversations/${conversationId}/files/content`,
-        config.dynamicAgentsUrl
-      );
-      backendUrl.searchParams.set("agent_id", agentId);
-      backendUrl.searchParams.set("path", path);
-
-      // Build headers for the backend request
-      const backendHeaders: Record<string, string> = {
-        "Content-Type": "application/json",
-      };
-
-      // Forward the access token to the Dynamic Agents backend
-      if (session.accessToken) {
-        backendHeaders["Authorization"] = `Bearer ${session.accessToken}`;
-      }
-
-      // Forward the request to the Dynamic Agents service
-      const response = await fetch(backendUrl.toString(), {
-        method: "DELETE",
-        headers: backendHeaders,
-      });
-
-      if (!response.ok) {
-        const errorText = await response.text();
-        let errorMessage = `Failed to delete file: ${response.status}`;
-        try {
-          const errorJson = JSON.parse(errorText);
-          errorMessage = errorJson.detail || errorMessage;
-        } catch {
-          // Use default error message
-        }
-        throw new ApiError(errorMessage, response.status);
-      }
-
-      const data = await response.json();
-      return NextResponse.json(data);
-    });
-  }
-);
+  return proxyRequest(resolved.backendUrl, "DELETE", resolved.authResult, "[files/content]");
+}
